+++ server/.env
@@ -1,22 +1,2 @@
 SECRET_KEY = f54e3a447c22a6055bffc622
-SQLALCHEMY_TRACK_MODIFICATIONS = False
-
-# Here you can globally configure all the ways you want to allow JWTs to
-# be sent to your web application. By default, this will be only headers.
-# app.config["JWT_TOKEN_LOCATION"] = ["headers", "cookies", "json", "query_string"]
-
-JWT_TOKEN_LOCATION = cookies
-
-# If true this will only allow the cookies that contain your JWTs to be sent
-# over https. In production, this should always be set to True
-JWT_COOKIE_SECURE = False
-
-# Change this in your code!
-JWT_SECRET_KEY= f54e3a447c22a6055bffc645
-
-
-GOOGLE_CLIENT_ID = 533430347161-rtfcp0m51ccd6d0rk0jqq8dr5tnfn28q.apps.googleusercontent.com
-GOOGLE_CLINET_SECRET = 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
-
-
-WEBSITE_URL = 'http://localhost:3000/'
+SQLALCHEMY_TRACK_MODIFICATIONS = false
